+++ src/GSvar/Background/BackgroundJobDialog.h
@@ -7,7 +7,6 @@
 #include <QThreadPool>
 #include <QDateTime>
 
-
 class BackgroundJobDialog
 	: public QDialog
 {
@@ -15,7 +14,9 @@ class BackgroundJobDialog
 
 public:
 	 BackgroundJobDialog(QWidget* parent);
-	 void start(BackgroundWorkerBase* job, bool show_busy_dialog);
+     int start(BackgroundWorkerBase* job, bool show_busy_dialog);
+     QString getJobStatus(int job_id);
+     QString getJobMessages(int job_id);
 
 private slots:
 	 void started();
@@ -26,18 +27,17 @@ private:
 	Ui::BackgroundJobDialog ui_;
 	QThreadPool pool_;
 	int next_id_;
+    struct JobInfo
+    {
+        int id = -1;
+        QString name;
+        QDateTime started;
+        QString status;
+        int elapsed_ms = -1;
+        QString messages;
 
-	struct JobInfo
-	{
-		int id = -1;
-		QString name;
-		QDateTime started;
-		QString status;
-		int elapsed_ms = -1;
-		QString messages;
-
-		BusyDialog* busy_dlg = nullptr;
-	};
+        BusyDialog* busy_dlg = nullptr;
+    };
 	QList<JobInfo> jobs_;
 
 	//Re-draws the table (or only one line, if id is given)